requirements.txt
python-dotenv
pandas
pydantic
MetaTrader5